module single signature - stmsig
