+++ mithril-stm/src/participant/signer.rs
@@ -1,7 +1,8 @@
 use crate::bls_multi_signature::{Signature, SigningKey, VerificationKey};
 use crate::eligibility_check::ev_lt_phi;
 use crate::key_reg::ClosedKeyReg;
-use crate::stm::{Stake, StmParameters, StmSig};
+use crate::single_signature::StmSig;
+use crate::stm::{Stake, StmParameters};
 use blake2::digest::{Digest, FixedOutput};
 
 /// Wrapper of the MultiSignature Verification key
