+++ mithril-stm/src/single_signature/mod.rs
@@ -1,0 +1,3 @@
+mod signature;
+
+pub use crate::single_signature::signature::StmSig;
